moved stats to blueprints
