+++ api/api/routes/stats.py
@@ -1,0 +1,34 @@
+from flask import Flask, request, session, send_from_directory, render_template
+from flask import Blueprint
+import api
+
+from api.common import WebSuccess, WebError
+from api.annotations import api_wrapper, require_login, require_teacher, require_admin, check_csrf
+from api.annotations import block_before_competition, block_after_competition
+from api.annotations import log_action
+
+blueprint = Blueprint("stats_api", __name__, template_folder="templates")
+
+@blueprint.route('/team/solved_problems', methods=['GET'])
+@api_wrapper
+@require_login
+@block_before_competition(WebError("The competition has not begun yet!"))
+def get_team_solved_problems_hook():
+    tid = request.args.get("tid", None)
+    stats = {
+        "problems": api.stats.get_problems_by_category(),
+        "members": api.stats.get_team_member_stats(tid)
+    }
+
+    return WebSuccess(data=stats)
+
+@blueprint.route('/team/score_progression', methods=['GET'])
+@api_wrapper
+@require_login
+@block_before_competition(WebError("The competition has not begun yet!"))
+def get_team_score_progression():
+    category = request.form.get("category", None)
+
+    tid = api.user.get_team()["tid"]
+
+    return WebSuccess(data=[api.stats.get_score_progression(tid=tid, category=category)])
